refactor(ts): split streamAgentChat.ts into focused modules (SRP)
The 532-line file violated SRP by mixing four concerns.  Extracted:

- wireMessages.ts     — AgentWireMessage types + convertToWireMessages()
- agentSseReader.ts   — readAgentSSE() async generator (54 lines)
- agentMessageState.ts — applyTextDelta / addToolCallPart /
                         applyToolResult / finalizeMessageTiming (144 lines)

streamAgentChat.ts is now the slim orchestrator (283 lines) that
imports from those three modules and handles only HTTP plumbing and
SSE event dispatch.
